still bootstrap, changed to domain driven design

+++ src/main/java/mk/trkalo/dnp/dnpshop/repository/ProductRepository.java
@@ -7,8 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     boolean existsByName(String name);
     List<Product> findAllByName(String name);
-    boolean existsProductByNameAndIdNot(String name, int id);
+    boolean existsProductByNameAndIdNot(String name, Long id);
+
 }
