+++ src/main/java/mk/trkalo/dnp/dnpshop/service/SizeService.java
@@ -5,7 +5,7 @@ import mk.trkalo.dnp.dnpshop.model.Size;
 import java.util.List;
 
 public interface SizeService {
-    Size findById(int id);
+    Size findById(Long id);
     List<Size> findAll();
     Size save(Size s);
     void deleteAll();
