+++ src/main/java/mk/trkalo/dnp/dnpshop/service/TypeService.java
@@ -5,7 +5,7 @@ import mk.trkalo.dnp.dnpshop.model.Type;
 import java.util.List;
 
 public interface TypeService {
-    Type findById(int id);
+    Type findById(Long id);
     List<Type> findAll();
     Type save(Type s);
     void deleteAll();
